requirements.txt
langgraph==0.1.19
langchain==0.2.6
langchain-google-genai==1.0.10
google-generativeai==0.7.2
python-dotenv==1.0.1
tavily-python==0.3.3
selenium==4.18.1
webdriver-manager==4.0.1
requests==2.31.0
gspread==6.0.2
oauth2client==4.1.3
email-validator==2.1.0
fastapi
uvicorn[standard]
pydantic